+++ maven-plugin/src/it/simple-run-with-report-path/verify.groovy
@@ -1,0 +1,3 @@
+File report = new File(basedir, "target/ijhttp-reports/report.xml")
+
+assert report.exists()