+++ spring-boot-test/src/it/autoconfiguration/src/main/java/uk/bot_by/ijhttp_tools/spring_boot_test/it.autoconfiguration/HttpClientCommandLineApplication.java
@@ -1,0 +1,13 @@
+package uk.bot_by.ijhttp_tools.spring_boot_test.it.autoconfiguration;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class HttpClientCommandLineApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(HttpClientCommandLineApplication.class, args);
+  }
+
+}