+++ spring-boot-test/src/it/autoconfiguration/verify.groovy
@@ -1,0 +1,3 @@
+File report = new File(basedir, "target/ijhttp/report.xml")
+
+assert report.file