+++ changelog.md
@@ -2,8 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## 1.1.0 - 2023-12-20
+### Added
+- The proxy feature
+- The report path
+- Split Maven plugin and CLI builder modules
+- The Spring Boot Test autoconfiguration
 
 ## 1.0.1 - 2023-11-08
 ### Fixed