+++ src/site/markdown/changelog.md
@@ -1,5 +1,21 @@
-# Change Log
+# Changelog
+
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/)
-and this project adheres to [Semantic Versioning](http://semver.org/).
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## 1.1.0 - 2023-12-20
+### Added
+- The proxy feature
+- The report path
+- Split Maven plugin and CLI builder modules
+- The Spring Boot Test autoconfiguration
+
+## 1.0.1 - 2023-11-08
+### Fixed
+- Environment variables, public and private, are added without handling quoting
+
+## 1.0.0 - 2023-05-05
+### Added
+- Executor Mojo: reads the configuration then runs `ijhttp`.